Move hashmap to .h file in impl directory.
